requirements/base.txt
# base requirements

Django==3.2.4
django-braces==1.14.0
django-taggit==1.4.0
psycopg2==2.9.1
django-oauth-toolkit==1.5.0
mistune==0.8.4
djangorestframework==3.12.4
redis==3.5.3
django-nested-admin==3.3.3
django-cors-headers==3.7.0
whitenoise==5.2.0
django-allauth==0.44.0
faker==8.8.1
